refactor(smt): configurable KVDB schema
The KVDB implementations must not depend on any particular schema.
The schema configuration is separated to KVDBSchemaFactory implementations.
Move related code there: new, configure_smt_cf_options, configure_block_table_options.
