+++ miden-crypto/src/merkle/smt/large/storage/fjall_kvdb.rs
@@ -6,10 +6,10 @@ use alloc::{boxed::Box, sync::Arc};
 use core::fmt;
 use std::collections::HashMap;
 
-use fjall::{Database, KeyspaceCreateOptions, OwnedWriteBatch, Readable, Snapshot};
+use fjall::{Database, OwnedWriteBatch, Readable, Snapshot};
 use lsm_tree;
 
-use super::{StorageError, StorageResult, config::*, kvdb::*, schema::*};
+use super::{StorageError, StorageResult, config::*, kvdb::*};
 
 impl From<fjall::Error> for StorageError {
     fn from(e: fjall::Error) -> Self {
@@ -29,54 +29,30 @@ impl fmt::Debug for FjallKVDB {
     }
 }
 
+pub struct FjallKVDBSchemaConfig {
+    pub db: Database,
+    pub keyspaces: HashMap<&'static str, fjall::Keyspace>,
+}
+
 impl KVDB for FjallKVDB {
     type Batch<'a> = FjallKVDBBatch<'a>;
+    type SchemaConfig = FjallKVDBSchemaConfig;
+    type SchemaConfigInput = Database;
 
-    fn new(config: PersistentSmtStorageConfig) -> StorageResult<Self> {
+    fn new(
+        config: PersistentSmtStorageConfig,
+        schema_factory: impl FnOnce(
+            PersistentSmtStorageConfig,
+            Database,
+        ) -> StorageResult<FjallKVDBSchemaConfig>,
+    ) -> StorageResult<Self> {
         let db = Database::builder(&config.path)
             .cache_size(config.cache_size as u64)
             .max_cached_files(Some(config.max_open_files as usize))
             .fsync_on_create(false)
             .open()?;
 
-        _ = ALL_TABLE_NAMES;
-
-        const SUBTREE_TABLES: &[&str] = &[
-            SUBTREE_16_CF,
-            SUBTREE_24_CF,
-            SUBTREE_32_CF,
-            SUBTREE_40_CF,
-            SUBTREE_48_CF,
-            SUBTREE_56_CF,
-        ];
-
-        let keyspace_common_opts =
-            || -> KeyspaceCreateOptions { KeyspaceCreateOptions::default().fsync_on_create(false) };
-
-        let mut keyspaces = HashMap::new();
-        for name in SUBTREE_TABLES {
-            let opts = keyspace_common_opts().max_memtable_size(128 << 20);
-            let keyspace = db.keyspace(name, || opts)?;
-            keyspaces.insert(*name, keyspace);
-        }
-
-        {
-            let opts = keyspace_common_opts().max_memtable_size(128 << 20);
-            let keyspace = db.keyspace(LEAVES_CF, || opts)?;
-            keyspaces.insert(LEAVES_CF, keyspace);
-        }
-
-        {
-            let opts = keyspace_common_opts();
-            let keyspace = db.keyspace(METADATA_CF, || opts)?;
-            keyspaces.insert(METADATA_CF, keyspace);
-        }
-
-        {
-            let opts = keyspace_common_opts();
-            let keyspace = db.keyspace(IN_MEM_DEPTH_CF, || opts)?;
-            keyspaces.insert(IN_MEM_DEPTH_CF, keyspace);
-        }
+        let FjallKVDBSchemaConfig { db, keyspaces } = schema_factory(config, db)?;
 
         if config.durability_mode == PersistentSmtStorageDurabilityMode::Relaxed {
             lsm_tree::file::set_fsync_enabled(false);
